Fix courses, and add user Resource view;

+++ components/resources/display/VideoItem.tsx
@@ -4,7 +4,7 @@ const VideoItem = ({url, onRemove=null}) => {
 
   const handleRemove = (e) => {
     e.preventDefault()
-    onRemove(e)
+    onRemove && onRemove(e)
   }
 
   return (
